always enable cli commands
